+++ sci/provider/datasource_applications.go
@@ -314,9 +314,8 @@ func (d *applicationsDataSource) Schema(_ context.Context, _ datasource.SchemaRe
 														Computed:            true,
 													},
 													"index": schema.Int32Attribute{
-														// TODO
-														// MarkdownDescription: "",
-														Computed: true,
+														MarkdownDescription: "A unique index for the endpoint.",
+														Computed:            true,
 													},
 													"default": schema.BoolAttribute{
 														MarkdownDescription: "Configure if the endpoint is the default one to be used.",
