+++ sci/provider/type_application.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 
 	"github.com/SAP/terraform-provider-sap-cloud-identity-services/internal/cli/apiObjects/applications"
 )
@@ -343,4 +344,150 @@ func applicationsValueFrom(ctx context.Context, a applications.ApplicationsRespo
 	}
 
 	return apps
+}
+
+// retrieve the API Request body from the plan data
+func getApplicationRequest(ctx context.Context, plan applicationData) (*applications.Application, diag.Diagnostics) {
+
+	var diagnostics, diags diag.Diagnostics
+
+	args := &applications.Application{
+		Name:           plan.Name.ValueString(),
+		Description:    plan.Description.ValueString(),
+		MultiTenantApp: plan.MultiTenantApp.ValueBool(),
+	}
+
+	if !plan.ParentApplicationId.IsNull() {
+		args.ParentApplicationId = plan.ParentApplicationId.ValueString()
+	}
+
+	// mapping of the plan data to the API request body must be done manually
+	// this is to ensure the proper handling of attributes where the API request body structure differs from that of the schema
+	// these attributes are : subject_name_identifier and advanced_assertion_attributes
+	// the schema defines them as objects with sub-attributes source and value, but the API request body expects them to be strings
+	if !plan.AuthenticationSchema.IsNull() && !plan.AuthenticationSchema.IsUnknown() {
+
+		args.AuthenticationSchema = &applications.AuthenticationSchema{}
+
+		var authenticationSchema authenticationSchemaData
+		diags = plan.AuthenticationSchema.As(ctx, &authenticationSchema, basetypes.ObjectAsOptions{
+			UnhandledNullAsEmpty:    true,
+			UnhandledUnknownAsEmpty: true,
+		})
+
+		diagnostics.Append(diags...)
+		if diagnostics.HasError() {
+			return nil, diagnostics
+		}
+
+		//SSO_TYPE
+		if !authenticationSchema.SsoType.IsNull() && !authenticationSchema.SsoType.IsUnknown() {
+			args.AuthenticationSchema.SsoType = authenticationSchema.SsoType.ValueString()
+		}
+
+		//DEFAULT_AUTHENTICATING_IDP
+		if !authenticationSchema.DefaultAuthenticatingIdpId.IsNull() && !authenticationSchema.DefaultAuthenticatingIdpId.IsUnknown() {
+			args.AuthenticationSchema.DefaultAuthenticatingIdpId = authenticationSchema.DefaultAuthenticatingIdpId.ValueString()
+		}
+
+		//SUBJECT_NAME_IDENTIFIER
+		if !authenticationSchema.SubjectNameIdentifier.IsNull() && !authenticationSchema.SubjectNameIdentifier.IsUnknown() {
+
+			var subjectNameIdentifier subjectNameIdentifierData
+			diags = authenticationSchema.SubjectNameIdentifier.As(ctx, &subjectNameIdentifier, basetypes.ObjectAsOptions{
+				UnhandledNullAsEmpty:    true,
+				UnhandledUnknownAsEmpty: true,
+			})
+
+			diagnostics.Append(diags...)
+			if diagnostics.HasError() {
+				return nil, diagnostics
+			}
+
+			// the mapping is done manually, in order to handle the parameter value when the source is set to "Corporate Identity Provider"
+			if subjectNameIdentifier.Source.ValueString() == sourceValues[0] || subjectNameIdentifier.Source.ValueString() == sourceValues[2] {
+				args.AuthenticationSchema.SubjectNameIdentifier = subjectNameIdentifier.Value.ValueString()
+			} else {
+				args.AuthenticationSchema.SubjectNameIdentifier = "${corporateIdP." + subjectNameIdentifier.Value.ValueString() + "}"
+			}
+		}
+
+		//SUBJECT_NAME_IDENTIFIER_FUNCTION
+		if !authenticationSchema.SubjectNameIdentifierFunction.IsNull() && !authenticationSchema.SubjectNameIdentifierFunction.IsUnknown() {
+			args.AuthenticationSchema.SubjectNameIdentifierFunction = authenticationSchema.SubjectNameIdentifierFunction.ValueString()
+		}
+
+		//ASSERTION_ATTRIBUTES
+		if !authenticationSchema.AssertionAttributes.IsNull() && !authenticationSchema.AssertionAttributes.IsUnknown() {
+
+			var attributes []applications.AssertionAttribute
+			diags := authenticationSchema.AssertionAttributes.ElementsAs(ctx, &attributes, true)
+			diagnostics.Append(diags...)
+			if diagnostics.HasError() {
+				return nil, diagnostics
+			}
+
+			args.AuthenticationSchema.AssertionAttributes = attributes
+		}
+
+		//ADVANCED_ASSERTION_ATTRIBUTES
+		if !authenticationSchema.AdvancedAssertionAttributes.IsNull() && !authenticationSchema.AdvancedAssertionAttributes.IsUnknown() {
+
+			var advancedAssertionAttributes []advancedAssertionAttributesData
+			diags := authenticationSchema.AdvancedAssertionAttributes.ElementsAs(ctx, &advancedAssertionAttributes, true)
+			diagnostics.Append(diags...)
+			if diagnostics.HasError() {
+				return nil, diagnostics
+			}
+
+			for _, attribute := range advancedAssertionAttributes {
+
+				assertionAttribute := applications.AdvancedAssertionAttribute{
+					AttributeName: attribute.AttributeName.ValueString(),
+				}
+
+				// the mapping is done manually, in order to handle the parameter attribute_value when the source is set to "Corporate Identity Provider"
+				if attribute.Source == types.StringValue(sourceValues[1]) {
+					assertionAttribute.AttributeValue = "${corporateIdP." + attribute.AttributeValue.ValueString() + "}"
+				} else {
+					assertionAttribute.AttributeValue = attribute.AttributeValue.ValueString()
+				}
+
+				args.AuthenticationSchema.AdvancedAssertionAttributes = append(args.AuthenticationSchema.AdvancedAssertionAttributes, assertionAttribute)
+			}
+		}
+
+		//AUTHENTICATION_RULES
+		if !authenticationSchema.AuthenticationRules.IsNull() {
+
+			var authrules []applications.AuthenicationRule
+			diags = authenticationSchema.AuthenticationRules.ElementsAs(ctx, &authrules, true)
+			diagnostics.Append(diags...)
+			if diagnostics.HasError() {
+				return nil, diagnostics
+			}
+
+			args.AuthenticationSchema.ConditionalAuthentication = authrules
+		}
+
+		//SAML2 CONFIGURATION
+		if !authenticationSchema.Saml2Configuration.IsNull() && !authenticationSchema.Saml2Configuration.IsUnknown() {
+
+			var saml2config applications.SamlConfiguration
+			diags := authenticationSchema.Saml2Configuration.As(ctx, &saml2config, basetypes.ObjectAsOptions{
+				UnhandledNullAsEmpty:    true,
+				UnhandledUnknownAsEmpty: true,
+			})
+			diagnostics.Append(diags...)
+
+			if diagnostics.HasError() {
+				return nil, diagnostics
+			}
+
+			args.AuthenticationSchema.Saml2Configuration = &saml2config
+		}
+
+	}
+
+	return args, diagnostics
 }
